+++ be/src/main/java/com/secondhand/domain/chat/dto/ChatRoom.java
@@ -1,19 +1,19 @@
-package com.secondhand.domain.chat;
+package com.secondhand.domain.chat.dto;
 
-import lombok.Getter;
-import lombok.Setter;
 
-import java.io.Serializable;
+import lombok.*;
+
 import java.util.UUID;
+
 
 @Getter
 @Setter
-public class ChatRoom implements Serializable {
-    private static final long serialVersionUID = 6494678977089006639L;
-
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class ChatRoom {
     private String roomId;
     private String name;
-    private long userCount; // 채팅방 인원수
 
     public static ChatRoom create(String name) {
         ChatRoom chatRoom = new ChatRoom();
@@ -22,3 +22,4 @@ public class ChatRoom implements Serializable {
         return chatRoom;
     }
 }
+
