--- be/src/main/java/com/secondhand/domain/chat/dto/request/ChatRequest.java
@@ -1,42 +1,0 @@
-package com.secondhand.domain.chat.dto.request;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-
-@Getter
-@NoArgsConstructor
-public class ChatRequest {
-    /**
-     * 송신자 id
-     */
-    @NotNull
-    private Long senderId;
-
-    /**
-     * 수신자 id
-     */
-    @NotNull
-    private Long receiverId;
-
-    /**
-     * 채팅방 id
-     */
-    @NotNull
-    private Long roomId;
-
-    /**
-     * 메시지 내용
-     */
-    @NotBlank
-    private String message;
-
-    public ChatRequest(Long senderId, Long receiverId, Long roomId, String message) {
-        this.senderId = senderId;
-        this.receiverId = receiverId;
-        this.roomId = roomId;
-        this.message = message;
-    }
-}
