+++ be/src/main/java/com/secondhand/service/ChatService.java
@@ -1,39 +1,52 @@
-package com.secondhand.service;
-
-import com.secondhand.domain.chat.repository.ChatRepository;
-import com.secondhand.domain.chat.dto.request.ChatRequest;
-import com.secondhand.domain.member.Member;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-public class ChatService {
-
-    private final ChatRepository chatRepository;
-
-    @Transactional
-    public void sendMessage(ChatRequest chatRequest, Member member) {
-//        ChatRoom chatRoom = chatRepository.findById(chatRequest.getRoomId()).orElseThrow(ChatRoomNotFoundException::new);
+//package com.secondhand.service;
 //
-//        //채팅 생성 및 저장
-//        ChatMessage chatMessage = ChatMessage.builder()
-//                .chatRoom(chatRoom)
-//                .user(member)
-//                .message(chatRequest.getMessage())
+//import com.fasterxml.jackson.databind.ObjectMapper;
+//import com.secondhand.domain.chat.ChatRoom;
+//import lombok.RequiredArgsConstructor;
+//import lombok.extern.slf4j.Slf4j;
+//import org.springframework.stereotype.Service;
+//import org.springframework.web.socket.TextMessage;
+//
+//import javax.annotation.PostConstruct;
+//import java.io.IOException;
+//import java.util.*;
+//
+//@Slf4j
+//@RequiredArgsConstructor
+//@Service
+//public class ChatService {
+//
+//    private final ObjectMapper objectMapper;
+//    private Map<String, ChatRoom> chatRooms;
+//
+//    @PostConstruct
+//    private void init() {
+//        chatRooms = new LinkedHashMap<>();
+//    }
+//
+//    public List<ChatRoom> findAllRoom() {
+//        return new ArrayList<>(chatRooms.values());
+//    }
+//
+//    public ChatRoom findRoomById(String roomId) {
+//        return chatRooms.get(roomId);
+//    }
+//
+//    public ChatRoom createRoom(String name) {
+//        String randomId = UUID.randomUUID().toString();
+//        ChatRoom chatRoom = ChatRoom.builder()
+//                .roomId(randomId)
+//                .name(name)
 //                .build();
+//        chatRooms.put(randomId, chatRoom);
+//        return chatRoom;
+//    }
 //
-//        chatMessageRepository.save(chatMessage);
-//        String topic = channelTopic.getTopic();
-//
-//        // ChatMessageRequest에 유저정보, 현재시간 저장
-//        chatRequest.setNickName(member.getNickname());
-//        chatRequest.setUserId(member.getId());
-//
-//        if (chatRequest.getType() == ChatMessageRequest.MessageType.TALK) {
-//            // 그륩 채팅일 경우
-//            redisTemplate.convertAndSend(topic, chatRequest);
+//    public <T> void sendMessage(WebSocketSession session, T message) {
+//        try {
+//            session.sendMessage(new TextMessage(objectMapper.writeValueAsString(message)));
+//        } catch (IOException e) {
+//            log.error(e.getMessage(), e);
 //        }
-    }
-}
+//    }
+//}
